hazard.msm {msm}	R Documentation

Calculate tables of hazard ratios for covariates on transition intensities

Description

Hazard ratios are computed by exponentiating the estimated covariate effects on the log-transition intensities. This function is called by summary.msm.

Usage

hazard.msm(x, hazard.scale = 1, cl = 0.95)

Arguments

x
Output from msm representing a fitted multi-state model.

hazard.scale
Vector with same elements as number of covariates on transition rates. Corresponds to the increase in each covariate used to calculate its hazard ratio. Defaults to all 1.

cl
Width of the symmetric confidence interval to present. Defaults to 0.95.

Value

A list of tables containing hazard ratio estimates, one table for each covariate. Each table has three columns, containing the hazard ratio, and an approximate upper and lower confidence limit respectively (assuming normality on the log scale), for each Markov chain transition intensity.

Author(s)

C. H. Jackson chris.jackson@mrc-bsu.cam.ac.uk

See Also

msm, summary.msm, odds.msm

[Package msm version 1.6.6 Index]