logLik.msm {msm}	R Documentation

Extract model log-likelihood

Description

Extract the log-likelihood and the number of parameters of a model fitted with msm.

Usage

## S3 method for class 'msm'
logLik(object, by.subject=FALSE, ...)

Arguments

object
A fitted multi-state model object, as returned by msm.

by.subject
Return vector of subject-specific log-likelihoods, which should sum to the total log-likelihood.

...
(unused) further arguments passed to or from other methods.

Value

The log-likelihood of the model represented by 'object' evaluated at the maximum likelihood estimates.

Akaike's information criterion can also be computed using AIC(object).

Author(s)

C. H. Jackson chris.jackson@mrc-bsu.cam.ac.uk

See Also

msm,lrtest.msm.

[Package msm version 1.6.6 Index]