print.msm {msm}	R Documentation

Print a fitted msm model object

Description

Print a fitted msm model object

Usage

## S3 method for class 'msm'
print(x,covariates=NULL, digits=4, ...)
printnew.msm(x, covariates=NULL, digits=4, ...)

Arguments

x
Output from msm, representing a fitted multi-state model object.

covariates
Covariates for which to print “baseline” transition intensities or misclassification probabilities. See qmatrix.msm for more details.

digits
Minimum number of significant digits, passed to format. Defaults to 4.

...
Other arguments to be passed to format.

Details

This is the new method of formatting msm objects for printing. The old method was based on printing lists of matrices. That produced a lot of wasted space for parameters which were zero, and it was difficult to match corresponding numbers between matrices. The new method presents all the transition intensities and covariate effects as a single compact table, and likewise for misclassification matrices.

Also in the old method, covariate effects were presented as log hazard ratios or log odds ratios. The log scale is more convenient mathematically, but unnatural to interpret. The new method presents hazard ratios for covariates on transition intensities and odds ratios for misclassification probabilities.

printnew.msm is an alias for print.msm.

Value

The object returned by print.msm is a numeric matrix with one column for each estimate or confidence limit for intensities and their covariates, in the same arrangement as printed, but with the underlying numbers in full precision. The results formatted for printing are stored in the "formatted" attribute of the object, as a character matrix. These can alternatively be produced by msm.form.qoutput, which has no printing side-effect. msm.form.eoutput produces the same arrangement for misclassification probabilities instead of intensities.

Author(s)

C. H. Jackson chris.jackson@mrc-bsu.cam.ac.uk

See Also

msm, printold.msm, msm.form.qoutput.

[Package msm version 1.6.6 Index]