qcmodel.object {msm}	R Documentation

Developer documentation: model for covariates on transition intensities

Description

A list representing the model for covariates on transition intensities

Value

npars
Number of covariate effect parameters. This is defined as the number of covariates on intensities (with factors expanded as contrasts) multiplied by the number of allowed transitions in the model.

Note if msm was called with covariates set to a list of different covariates for different intensities, then this will include covariate effects that are implicitly defined as zero by this list. The information in paramdata objects can be used to identify wich ones are fixed at zero.

This also includes any timeperiod covariates in a time-inhomogeneous model defined by the pci option to msm.

ndpars
Number of distinct covariate effect parameters, as npars, but after any equality constraints have been applied.

ncovs
Number of covariates on intensities, with factors expanded as contrasts.

constr
List of equality constraints on these covariate effects, as supplied in the constraint argument to msm.

covlabels
Names / labels of these covariates in the model matrix (see model.matrix.msm).

inits
Initial values for these covariate effects, as a vector formed from the covinits list supplied to msm.

covmeans
Means of these covariates in the data (excluding data not required to fit the model, such as observations with missing data in other elements or subjects' last observations). This includes means of 0/1 factor contrasts as well as continuous covariates (for historic reasons, which may not be sensible).

See Also

msm.object.

[Package msm version 1.6.6 Index]