recreate.olddata {msm}	R Documentation

Convert data stored in msm object to old format

Description

Converts the data element of msm objects to the old format.

Usage

recreate.olddata(x)

Arguments

x
Object returned by the msm function, representing a fitted multi-state model.

Details

This is just provided for convenience and to illustrate the changes. It is not guaranteed to be complete, and is liable to be withdrawn. Users who were relying on the previous undocumented format are advised to upgrade their code to use the new format, which uses model frames and model design matrices in the standard format used in version 1.4, based on model.frame and model.matrix.

Value

A list of vectors and matrices in the undocumented ad-hoc format used for the data component of msm objects in msm versions 1.3.1 and earlier.

[Package msm version 1.6.6 Index]