msm.summary {msm}	R Documentation

Summarise a fitted multi-state model

Description

Summary method for fitted msm models. This is simply a wrapper around prevalence.msm which produces a table of observed and expected state prevalences for each time, and for models with covariates, hazard.msm to print hazard ratios with 95% confidence intervals for covariate effects.

Usage

## S3 method for class 'msm'
summary(object, hazard.scale=1, ...)

Arguments

object
A fitted multi-state model object, as returned by msm.

hazard.scale
Vector with same elements as number of covariates on transition rates. Corresponds to the increase in each covariate used to calculate its hazard ratio. Defaults to all 1.

...
Further arguments passed to prevalence.msm.

Value

A list of class summary.msm, with components:

prevalences
Output from prevalence.msm.

hazard
Output from hazard.msm.

hazard.scale
Value of the hazard.scale argument.

Author(s)

C. H. Jackson chris.jackson@mrc-bsu.cam.ac.uk

See Also

msm,prevalence.msm, hazard.msm

[Package msm version 1.6.6 Index]