tidy.msm.estbystate {msm}	R Documentation

Tidy the output of totlos.msm and similar functions

Description

Note this should be called as tidy() not tidy.msm.totlos() or anything else, as this is a method for the generic tidy() function.

Usage

## S3 method for class 'msm.estbystate'
tidy(x, ...)

Arguments

x
Output of totlos.msm, envisits.msm or efpt.msm, which return objects of class "msm.estbystate".

...
Further arguments (unused).

Value

A tibble with one row per state, and columns for the estimate, and confidence intervals if available.

[Package msm version 1.8.2 Index]