transient.msm {msm}	R Documentation

Transient and absorbing states

Description

Returns the transient and absorbing states of either a fitted model or a transition intensity matrix.

Usage

transient.msm(x=NULL, qmatrix=NULL)
absorbing.msm(x=NULL, qmatrix=NULL)

Arguments

x
A fitted multi-state model as returned by msm.

qmatrix
A transition intensity matrix. The diagonal is ignored and taken to be minus the sum of the rest of the row.

Value

A vector of the ordinal indices of the transient or absorbing states.

Author(s)

C. H. Jackson chris.jackson@mrc-bsu.cam.ac.uk

[Package msm version 1.6.6 Index]