Multi-State Markov and Hidden Markov Models in Continuous Time

msm: Functions for fitting continuous-time Markov and hidden Markov multi-state models to longitudinal data. Designed for processes observed at arbitrary times in continuous time (intermittently observed or panel data) but some other observation schemes are supported. Both Markov transition rates and the hidden Markov output process can be modelled in terms of covariates, which may be constant or piecewise-constant in time.

See also

Author

Maintainer: Christopher Jackson chris.jackson@mrc-bsu.cam.ac.uk