Fitted msm model objects

The msm function returns a list with the following components. These are intended for developers and confident users. To extract results from fitted model objects, functions such as qmatrix.msm or print.msm should be used instead.

Value

call

The original call to msm, as returned by match.call.

Qmatrices

A list of matrices. The first component, labelled logbaseline, is a matrix containing the estimated transition intensities on the log scale with any covariates fixed at their means in the data (or at zero, if center=FALSE). The component labelled baseline is the equivalent on the untransformed scale. Each remaining component is a matrix giving the linear effects of the labelled covariate on the matrix of log intensities. To extract an estimated intensity matrix on the natural scale, at an arbitrary combination of covariate values, use the function qmatrix.msm.

QmatricesSE

The standard error matrices corresponding to Qmatrices.

QmatricesL,QmatricesU

Corresponding lower and upper symmetric confidence limits, of width 0.95 unless specified otherwise by the cl argument.

Ematrices

A list of matrices. The first component, labelled logitbaseline, is the estimated misclassification probability matrix (expressed as as log odds relative to the probability of the true state) with any covariates fixed at their means in the data (or at zero, if center=FALSE). The component labelled baseline is the equivalent on the untransformed scale. Each remaining component is a matrix giving the linear effects of the labelled covariate on the matrix of logit misclassification probabilities. To extract an estimated misclassification probability matrix on the natural scale, at an arbitrary combination of covariate values, use the function ematrix.msm.

EmatricesSE

The standard error matrices corresponding to Ematrices.

EmatricesL,EmatricesU

Corresponding lower and upper symmetric confidence limits, of width 0.95 unless specified otherwise by the cl argument.

minus2loglik

Minus twice the maximised log-likelihood.

deriv

Derivatives of the minus twice log-likelihood at its maximum.

estimates

Vector of untransformed maximum likelihood estimates returned from optim. Transition intensities are on the log scale and misclassification probabilities are given as log odds relative to the probability of the true state.

estimates.t

Vector of transformed maximum likelihood estimates with intensities and probabilities on their natural scales.

fixedpars

Indices of estimates which were fixed during the maximum likelihood estimation.

center

Indicator for whether the estimation was performed with covariates centered on their means in the data.

covmat

Covariance matrix corresponding to estimates.

ci

Matrix of confidence intervals corresponding to estimates.t

opt

Return value from the optimisation routine (such as optim or nlm), giving information about the results of the optimisation.

foundse

Logical value indicating whether the Hessian was positive-definite at the supposed maximum of the likelihood. If not, the covariance matrix of the parameters is unavailable. In these cases the optimisation has probably not converged to a maximum.

data

A list giving the data used for the model fit, for use in post-processing. To extract it, use the methods model.frame.msm or model.matrix.msm.

The format of this element changed in version 1.4 of msm, so that it now contains a model.frame object mf with all the variables used in the model. The previous format (an ad-hoc list of vectors and matrices) can be obtained with the function recreate.olddata(msmobject), where msmobject is the object returned by msm.

qmodel

A list of objects representing the transition matrix structure and options for likelihood calculation. See qmodel.object for documentation of the components.

emodel

A list of objects representing the misclassification model structure, for models specified using the ematrix argument to msm. See emodel.object.

qcmodel

A list of objects representing the model for covariates on transition intensities. See qcmodel.object.

ecmodel

A list of objects representing the model for covariates on transition intensities. See ecmodel.object.

hmodel

A list of objects representing the hidden Markov model structure. See hmodel.object.

cmodel

A list giving information about censored states. See cmodel.object.

pci

Cut points for time-varying intensities, as supplied to msm, but excluding any that are outside the times observed in the data.

paramdata

A list giving information about the parameters of the multi-state model. See paramdata.object.

cl

Confidence interval width, as supplied to msm.

covariates

Formula for covariates on intensities, as supplied to msm.

misccovariates

Formula for covariates on misclassification probabilities, as supplied to msm.

hcovariates

Formula for covariates on hidden Markov model outcomes, as supplied to msm.

initcovariates

Formula for covariates on initial state occupancy probabilities in hidden Markov models, as supplied to msm.

sojourn

A list as returned by sojourn.msm, with components:

mean = estimated mean sojourn times in the transient states, with covariates fixed at their means (if center=TRUE) or at zero (if center=FALSE).

se = corresponding standard errors.