Simulate multiple trajectories from a multi-state Markov model with arbitrary observation times

Simulate a number of individual realisations from a continuous-time Markov process. Observations of the process are made at specified arbitrary times for each individual, giving panel-observed data.

Usage

simmulti.msm(
  data,
  qmatrix,
  covariates = NULL,
  death = FALSE,
  start,
  ematrix = NULL,
  misccovariates = NULL,
  hmodel = NULL,
  hcovariates = NULL,
  censor.states = NULL,
  drop.absorb = TRUE
)

Arguments

data

A data frame with a mandatory column named time, representing observation times. The optional column named subject, corresponds to subject identification numbers. If not given, all observations are assumed to be on the same individual. Observation times should be sorted within individuals. The optional column named cens indicates the times at which simulated states should be censored. If cens==0 then the state is not censored, and if cens==k, say, then all simulated states at that time which are in the set censor.states are replaced by k. Other named columns of the data frame represent any covariates, which may be time-constant or time-dependent. Time-dependent covariates are assumed to be constant between the observation times.

qmatrix

The transition intensity matrix of the Markov process, with any covariates set to zero. The diagonal of qmatrix is ignored, and computed as appropriate so that the rows sum to zero. For example, a possible qmatrix for a three state illness-death model with recovery is:

rbind( c( 0, 0.1, 0.02 ), c( 0.1, 0, 0.01 ), c( 0, 0, 0 ) )

covariates

List of linear covariate effects on log transition intensities. Each element is a vector of the effects of one covariate on all the transition intensities. The intensities are ordered by reading across rows of the intensity matrix, starting with the first, counting the positive off-diagonal elements of the matrix.

For example, for a multi-state model with three transition intensities, and two covariates x and y on each intensity,

covariates=list(x = c(-0.3,-0.3,-0.3), y=c(0.1, 0.1, 0.1))

death

Vector of indices of the death states. A death state is an absorbing state whose time of entry is known exactly, but the individual is assumed to be in an unknown transient state ("alive") at the previous instant. This is the usual situation for times of death in chronic disease monitoring data. For example, if you specify death = c(4, 5) then states 4 and 5 are assumed to be death states.

death = TRUE indicates that the final state is a death state, and death = FALSE (the default) indicates that there is no death state.

start

A vector with the same number of elements as there are distinct subjects in the data, giving the states in which each corresponding individual begins. Or a single number, if all of these are the same. Defaults to state 1 for each subject.

ematrix

An optional misclassification matrix for generating observed states conditionally on the simulated true states. As defined in msm.

misccovariates

Covariate effects on misclassification probabilities via multinomial logistic regression. Linear effects operate on the log of each probability relative to the probability of classification in the correct state. In same format as covariates.

hmodel

An optional hidden Markov model for generating observed outcomes conditionally on the simulated true states. As defined in msm. Multivariate outcomes (hmmMV) are not supported.

hcovariates

List of the same length as hmodel, defining any covariates governing the hidden Markov outcome models. Unlike in the msm function, this should also define the values of the covariate effects. Each element of the list is a named vector of the initial values for each set of covariates for that state. For example, for a three-state hidden Markov model with two, one and no covariates on the state 1, 2 and 3 outcome models respectively,

hcovariates = list (c(acute=-8, age=0), c(acute=-8), NULL)

censor.states

Set of simulated states which should be replaced by a censoring indicator at censoring times. By default this is all transient states (representing alive, with unknown state).

drop.absorb

Drop repeated observations in the absorbing state, retaining only one.

Value

A data frame with columns,

subject

Subject identification indicators

time

Observation times

state

Simulated (true) state at the corresponding time

obs

Observed outcome at the corresponding time, if ematrix or hmodel was supplied

keep

Row numbers of the original data. Useful when drop.absorb=TRUE, to show which rows were not dropped

plus any supplied covariates.

Details

sim.msm is called repeatedly to produce a simulated trajectory for each individual. The state at each specified observation time is then taken to produce a new column state. The effect of time-dependent covariates on the transition intensity matrix for an individual is determined by assuming that the covariate is a step function which remains constant in between the individual's observation times. If the subject enters an absorbing state, then only the first observation in that state is kept in the data frame. Rows corresponding to future observations are deleted. The entry times into states given in death are assumed to be known exactly.

See also

Author

C. H. Jackson chris.jackson@mrc-bsu.cam.ac.uk

Examples

### Simulate 100 individuals with common observation times
sim.df <- data.frame(subject = rep(1:100, rep(13,100)), time = rep(seq(0, 24, 2), 100))
qmatrix <- rbind(c(-0.11,   0.1,  0.01 ),
                 c(0.05,   -0.15,  0.1 ),
                 c(0.02,   0.07, -0.09))
simmulti.msm(sim.df, qmatrix)
#>      subject time state keep
#> 1          1    0     1    1
#> 2          1    2     1    2
#> 3          1    4     1    3
#> 4          1    6     2    4
#> 5          1    8     2    5
#> 6          1   10     2    6
#> 7          1   12     2    7
#> 8          1   14     3    8
#> 9          1   16     3    9
#> 10         1   18     2   10
#> 11         1   20     3   11
#> 12         1   22     3   12
#> 13         1   24     3   13
#> 14         2    0     1   14
#> 15         2    2     3   15
#> 16         2    4     3   16
#> 17         2    6     2   17
#> 18         2    8     2   18
#> 19         2   10     2   19
#> 20         2   12     3   20
#> 21         2   14     3   21
#> 22         2   16     3   22
#> 23         2   18     3   23
#> 24         2   20     2   24
#> 25         2   22     3   25
#> 26         2   24     3   26
#> 27         3    0     1   27
#> 28         3    2     1   28
#> 29         3    4     2   29
#> 30         3    6     3   30
#> 31         3    8     3   31
#> 32         3   10     3   32
#> 33         3   12     3   33
#> 34         3   14     3   34
#> 35         3   16     3   35
#> 36         3   18     3   36
#> 37         3   20     3   37
#> 38         3   22     3   38
#> 39         3   24     3   39
#> 40         4    0     1   40
#> 41         4    2     2   41
#> 42         4    4     2   42
#> 43         4    6     2   43
#> 44         4    8     2   44
#> 45         4   10     2   45
#> 46         4   12     2   46
#> 47         4   14     2   47
#> 48         4   16     2   48
#> 49         4   18     2   49
#> 50         4   20     2   50
#> 51         4   22     2   51
#> 52         4   24     1   52
#> 53         5    0     1   53
#> 54         5    2     1   54
#> 55         5    4     1   55
#> 56         5    6     2   56
#> 57         5    8     2   57
#> 58         5   10     2   58
#> 59         5   12     2   59
#> 60         5   14     2   60
#> 61         5   16     2   61
#> 62         5   18     2   62
#> 63         5   20     2   63
#> 64         5   22     2   64
#> 65         5   24     2   65
#> 66         6    0     1   66
#> 67         6    2     2   67
#> 68         6    4     2   68
#> 69         6    6     3   69
#> 70         6    8     3   70
#> 71         6   10     3   71
#> 72         6   12     3   72
#> 73         6   14     3   73
#> 74         6   16     3   74
#> 75         6   18     3   75
#> 76         6   20     2   76
#> 77         6   22     2   77
#> 78         6   24     3   78
#> 79         7    0     1   79
#> 80         7    2     2   80
#> 81         7    4     2   81
#> 82         7    6     2   82
#> 83         7    8     2   83
#> 84         7   10     2   84
#> 85         7   12     2   85
#> 86         7   14     3   86
#> 87         7   16     3   87
#> 88         7   18     3   88
#> 89         7   20     3   89
#> 90         7   22     3   90
#> 91         7   24     3   91
#> 92         8    0     1   92
#> 93         8    2     1   93
#> 94         8    4     1   94
#> 95         8    6     1   95
#> 96         8    8     1   96
#> 97         8   10     1   97
#> 98         8   12     1   98
#> 99         8   14     1   99
#> 100        8   16     2  100
#> 101        8   18     2  101
#> 102        8   20     2  102
#> 103        8   22     2  103
#> 104        8   24     2  104
#> 105        9    0     1  105
#> 106        9    2     1  106
#> 107        9    4     1  107
#> 108        9    6     1  108
#> 109        9    8     1  109
#> 110        9   10     1  110
#> 111        9   12     1  111
#> 112        9   14     1  112
#> 113        9   16     2  113
#> 114        9   18     2  114
#> 115        9   20     1  115
#> 116        9   22     1  116
#> 117        9   24     2  117
#> 118       10    0     1  118
#> 119       10    2     2  119
#> 120       10    4     3  120
#> 121       10    6     2  121
#> 122       10    8     2  122
#> 123       10   10     3  123
#> 124       10   12     3  124
#> 125       10   14     2  125
#> 126       10   16     2  126
#> 127       10   18     2  127
#> 128       10   20     2  128
#> 129       10   22     2  129
#> 130       10   24     2  130
#> 131       11    0     1  131
#> 132       11    2     2  132
#> 133       11    4     2  133
#> 134       11    6     1  134
#> 135       11    8     1  135
#> 136       11   10     1  136
#> 137       11   12     1  137
#> 138       11   14     1  138
#> 139       11   16     3  139
#> 140       11   18     3  140
#> 141       11   20     3  141
#> 142       11   22     3  142
#> 143       11   24     3  143
#> 144       12    0     1  144
#> 145       12    2     1  145
#> 146       12    4     1  146
#> 147       12    6     1  147
#> 148       12    8     1  148
#> 149       12   10     1  149
#> 150       12   12     2  150
#> 151       12   14     2  151
#> 152       12   16     2  152
#> 153       12   18     1  153
#> 154       12   20     1  154
#> 155       12   22     1  155
#> 156       12   24     1  156
#> 157       13    0     1  157
#> 158       13    2     2  158
#> 159       13    4     2  159
#> 160       13    6     2  160
#> 161       13    8     2  161
#> 162       13   10     3  162
#> 163       13   12     3  163
#> 164       13   14     3  164
#> 165       13   16     2  165
#> 166       13   18     2  166
#> 167       13   20     2  167
#> 168       13   22     1  168
#> 169       13   24     1  169
#> 170       14    0     1  170
#> 171       14    2     1  171
#> 172       14    4     1  172
#> 173       14    6     1  173
#> 174       14    8     2  174
#> 175       14   10     2  175
#> 176       14   12     2  176
#> 177       14   14     2  177
#> 178       14   16     3  178
#> 179       14   18     3  179
#> 180       14   20     3  180
#> 181       14   22     3  181
#> 182       14   24     3  182
#> 183       15    0     1  183
#> 184       15    2     1  184
#> 185       15    4     2  185
#> 186       15    6     2  186
#> 187       15    8     2  187
#> 188       15   10     2  188
#> 189       15   12     2  189
#> 190       15   14     2  190
#> 191       15   16     3  191
#> 192       15   18     3  192
#> 193       15   20     3  193
#> 194       15   22     3  194
#> 195       15   24     3  195
#> 196       16    0     1  196
#> 197       16    2     1  197
#> 198       16    4     2  198
#> 199       16    6     1  199
#> 200       16    8     1  200
#> 201       16   10     1  201
#> 202       16   12     1  202
#> 203       16   14     1  203
#> 204       16   16     1  204
#> 205       16   18     2  205
#> 206       16   20     2  206
#> 207       16   22     2  207
#> 208       16   24     2  208
#> 209       17    0     1  209
#> 210       17    2     1  210
#> 211       17    4     1  211
#> 212       17    6     1  212
#> 213       17    8     1  213
#> 214       17   10     1  214
#> 215       17   12     1  215
#> 216       17   14     1  216
#> 217       17   16     1  217
#> 218       17   18     1  218
#> 219       17   20     1  219
#> 220       17   22     1  220
#> 221       17   24     1  221
#> 222       18    0     1  222
#> 223       18    2     1  223
#> 224       18    4     1  224
#> 225       18    6     1  225
#> 226       18    8     1  226
#> 227       18   10     1  227
#> 228       18   12     2  228
#> 229       18   14     1  229
#> 230       18   16     1  230
#> 231       18   18     1  231
#> 232       18   20     1  232
#> 233       18   22     1  233
#> 234       18   24     1  234
#> 235       19    0     1  235
#> 236       19    2     2  236
#> 237       19    4     2  237
#> 238       19    6     2  238
#> 239       19    8     3  239
#> 240       19   10     3  240
#> 241       19   12     1  241
#> 242       19   14     1  242
#> 243       19   16     1  243
#> 244       19   18     1  244
#> 245       19   20     1  245
#> 246       19   22     3  246
#> 247       19   24     3  247
#> 248       20    0     1  248
#> 249       20    2     1  249
#> 250       20    4     1  250
#> 251       20    6     2  251
#> 252       20    8     3  252
#> 253       20   10     3  253
#> 254       20   12     3  254
#> 255       20   14     3  255
#> 256       20   16     3  256
#> 257       20   18     3  257
#> 258       20   20     3  258
#> 259       20   22     3  259
#> 260       20   24     3  260
#> 261       21    0     1  261
#> 262       21    2     1  262
#> 263       21    4     1  263
#> 264       21    6     1  264
#> 265       21    8     1  265
#> 266       21   10     1  266
#> 267       21   12     1  267
#> 268       21   14     1  268
#> 269       21   16     1  269
#> 270       21   18     1  270
#> 271       21   20     1  271
#> 272       21   22     1  272
#> 273       21   24     2  273
#> 274       22    0     1  274
#> 275       22    2     2  275
#> 276       22    4     2  276
#> 277       22    6     2  277
#> 278       22    8     2  278
#> 279       22   10     2  279
#> 280       22   12     2  280
#> 281       22   14     2  281
#> 282       22   16     2  282
#> 283       22   18     2  283
#> 284       22   20     2  284
#> 285       22   22     2  285
#> 286       22   24     2  286
#> 287       23    0     1  287
#> 288       23    2     1  288
#> 289       23    4     1  289
#> 290       23    6     1  290
#> 291       23    8     2  291
#> 292       23   10     2  292
#> 293       23   12     3  293
#> 294       23   14     3  294
#> 295       23   16     3  295
#> 296       23   18     3  296
#> 297       23   20     3  297
#> 298       23   22     3  298
#> 299       23   24     3  299
#> 300       24    0     1  300
#> 301       24    2     1  301
#> 302       24    4     1  302
#> 303       24    6     1  303
#> 304       24    8     1  304
#> 305       24   10     1  305
#> 306       24   12     1  306
#> 307       24   14     1  307
#> 308       24   16     1  308
#> 309       24   18     1  309
#> 310       24   20     1  310
#> 311       24   22     1  311
#> 312       24   24     1  312
#> 313       25    0     1  313
#> 314       25    2     1  314
#> 315       25    4     1  315
#> 316       25    6     1  316
#> 317       25    8     1  317
#> 318       25   10     1  318
#> 319       25   12     1  319
#> 320       25   14     1  320
#> 321       25   16     1  321
#> 322       25   18     1  322
#> 323       25   20     1  323
#> 324       25   22     1  324
#> 325       25   24     1  325
#> 326       26    0     1  326
#> 327       26    2     1  327
#> 328       26    4     1  328
#> 329       26    6     1  329
#> 330       26    8     1  330
#> 331       26   10     1  331
#> 332       26   12     1  332
#> 333       26   14     1  333
#> 334       26   16     2  334
#> 335       26   18     1  335
#> 336       26   20     1  336
#> 337       26   22     1  337
#> 338       26   24     1  338
#> 339       27    0     1  339
#> 340       27    2     2  340
#> 341       27    4     2  341
#> 342       27    6     3  342
#> 343       27    8     3  343
#> 344       27   10     3  344
#> 345       27   12     3  345
#> 346       27   14     3  346
#> 347       27   16     3  347
#> 348       27   18     3  348
#> 349       27   20     2  349
#> 350       27   22     2  350
#> 351       27   24     1  351
#> 352       28    0     1  352
#> 353       28    2     1  353
#> 354       28    4     1  354
#> 355       28    6     1  355
#> 356       28    8     1  356
#> 357       28   10     3  357
#> 358       28   12     3  358
#> 359       28   14     3  359
#> 360       28   16     3  360
#> 361       28   18     3  361
#> 362       28   20     3  362
#> 363       28   22     3  363
#> 364       28   24     2  364
#> 365       29    0     1  365
#> 366       29    2     1  366
#> 367       29    4     1  367
#> 368       29    6     2  368
#> 369       29    8     2  369
#> 370       29   10     2  370
#> 371       29   12     2  371
#> 372       29   14     2  372
#> 373       29   16     2  373
#> 374       29   18     3  374
#> 375       29   20     1  375
#> 376       29   22     1  376
#> 377       29   24     1  377
#> 378       30    0     1  378
#> 379       30    2     1  379
#> 380       30    4     1  380
#> 381       30    6     1  381
#> 382       30    8     1  382
#> 383       30   10     1  383
#> 384       30   12     1  384
#> 385       30   14     1  385
#> 386       30   16     1  386
#> 387       30   18     1  387
#> 388       30   20     1  388
#> 389       30   22     1  389
#> 390       30   24     1  390
#> 391       31    0     1  391
#> 392       31    2     1  392
#> 393       31    4     1  393
#> 394       31    6     1  394
#> 395       31    8     2  395
#> 396       31   10     2  396
#> 397       31   12     2  397
#> 398       31   14     2  398
#> 399       31   16     2  399
#> 400       31   18     1  400
#> 401       31   20     1  401
#> 402       31   22     1  402
#> 403       31   24     3  403
#> 404       32    0     1  404
#> 405       32    2     1  405
#> 406       32    4     1  406
#> 407       32    6     2  407
#> 408       32    8     2  408
#> 409       32   10     2  409
#> 410       32   12     2  410
#> 411       32   14     2  411
#> 412       32   16     2  412
#> 413       32   18     1  413
#> 414       32   20     1  414
#> 415       32   22     1  415
#> 416       32   24     1  416
#> 417       33    0     1  417
#> 418       33    2     1  418
#> 419       33    4     1  419
#> 420       33    6     1  420
#> 421       33    8     1  421
#> 422       33   10     2  422
#> 423       33   12     2  423
#> 424       33   14     2  424
#> 425       33   16     3  425
#> 426       33   18     3  426
#> 427       33   20     3  427
#> 428       33   22     2  428
#> 429       33   24     2  429
#> 430       34    0     1  430
#> 431       34    2     1  431
#> 432       34    4     1  432
#> 433       34    6     1  433
#> 434       34    8     2  434
#> 435       34   10     2  435
#> 436       34   12     1  436
#> 437       34   14     2  437
#> 438       34   16     3  438
#> 439       34   18     3  439
#> 440       34   20     3  440
#> 441       34   22     3  441
#> 442       34   24     2  442
#> 443       35    0     1  443
#> 444       35    2     1  444
#> 445       35    4     1  445
#> 446       35    6     1  446
#> 447       35    8     2  447
#> 448       35   10     3  448
#> 449       35   12     3  449
#> 450       35   14     3  450
#> 451       35   16     3  451
#> 452       35   18     3  452
#> 453       35   20     2  453
#> 454       35   22     3  454
#> 455       35   24     3  455
#> 456       36    0     1  456
#> 457       36    2     1  457
#> 458       36    4     1  458
#> 459       36    6     1  459
#> 460       36    8     1  460
#> 461       36   10     1  461
#> 462       36   12     1  462
#> 463       36   14     1  463
#> 464       36   16     1  464
#> 465       36   18     1  465
#> 466       36   20     1  466
#> 467       36   22     1  467
#> 468       36   24     1  468
#> 469       37    0     1  469
#> 470       37    2     1  470
#> 471       37    4     1  471
#> 472       37    6     1  472
#> 473       37    8     1  473
#> 474       37   10     1  474
#> 475       37   12     1  475
#> 476       37   14     1  476
#> 477       37   16     3  477
#> 478       37   18     3  478
#> 479       37   20     3  479
#> 480       37   22     3  480
#> 481       37   24     3  481
#> 482       38    0     1  482
#> 483       38    2     1  483
#> 484       38    4     1  484
#> 485       38    6     1  485
#> 486       38    8     1  486
#> 487       38   10     2  487
#> 488       38   12     2  488
#> 489       38   14     2  489
#> 490       38   16     2  490
#> 491       38   18     2  491
#> 492       38   20     3  492
#> 493       38   22     3  493
#> 494       38   24     3  494
#> 495       39    0     1  495
#> 496       39    2     1  496
#> 497       39    4     1  497
#> 498       39    6     1  498
#> 499       39    8     1  499
#> 500       39   10     2  500
#> 501       39   12     2  501
#> 502       39   14     2  502
#> 503       39   16     2  503
#> 504       39   18     2  504
#> 505       39   20     2  505
#> 506       39   22     2  506
#> 507       39   24     2  507
#> 508       40    0     1  508
#> 509       40    2     1  509
#> 510       40    4     1  510
#> 511       40    6     1  511
#> 512       40    8     1  512
#> 513       40   10     1  513
#> 514       40   12     2  514
#> 515       40   14     2  515
#> 516       40   16     2  516
#> 517       40   18     2  517
#> 518       40   20     2  518
#> 519       40   22     2  519
#> 520       40   24     2  520
#> 521       41    0     1  521
#> 522       41    2     1  522
#> 523       41    4     1  523
#> 524       41    6     2  524
#> 525       41    8     2  525
#> 526       41   10     2  526
#> 527       41   12     2  527
#> 528       41   14     2  528
#> 529       41   16     2  529
#> 530       41   18     2  530
#> 531       41   20     2  531
#> 532       41   22     3  532
#> 533       41   24     3  533
#> 534       42    0     1  534
#> 535       42    2     1  535
#> 536       42    4     1  536
#> 537       42    6     1  537
#> 538       42    8     1  538
#> 539       42   10     2  539
#> 540       42   12     3  540
#> 541       42   14     2  541
#> 542       42   16     3  542
#> 543       42   18     3  543
#> 544       42   20     3  544
#> 545       42   22     3  545
#> 546       42   24     3  546
#> 547       43    0     1  547
#> 548       43    2     2  548
#> 549       43    4     2  549
#> 550       43    6     2  550
#> 551       43    8     2  551
#> 552       43   10     2  552
#> 553       43   12     2  553
#> 554       43   14     2  554
#> 555       43   16     2  555
#> 556       43   18     2  556
#> 557       43   20     2  557
#> 558       43   22     2  558
#> 559       43   24     2  559
#> 560       44    0     1  560
#> 561       44    2     1  561
#> 562       44    4     1  562
#> 563       44    6     2  563
#> 564       44    8     2  564
#> 565       44   10     2  565
#> 566       44   12     1  566
#> 567       44   14     1  567
#> 568       44   16     1  568
#> 569       44   18     1  569
#> 570       44   20     2  570
#> 571       44   22     2  571
#> 572       44   24     2  572
#> 573       45    0     1  573
#> 574       45    2     1  574
#> 575       45    4     1  575
#> 576       45    6     2  576
#> 577       45    8     2  577
#> 578       45   10     2  578
#> 579       45   12     3  579
#> 580       45   14     3  580
#> 581       45   16     3  581
#> 582       45   18     3  582
#> 583       45   20     3  583
#> 584       45   22     3  584
#> 585       45   24     3  585
#> 586       46    0     1  586
#> 587       46    2     1  587
#> 588       46    4     1  588
#> 589       46    6     2  589
#> 590       46    8     2  590
#> 591       46   10     2  591
#> 592       46   12     2  592
#> 593       46   14     2  593
#> 594       46   16     2  594
#> 595       46   18     3  595
#> 596       46   20     3  596
#> 597       46   22     3  597
#> 598       46   24     2  598
#> 599       47    0     1  599
#> 600       47    2     1  600
#> 601       47    4     1  601
#> 602       47    6     1  602
#> 603       47    8     1  603
#> 604       47   10     2  604
#> 605       47   12     2  605
#> 606       47   14     3  606
#> 607       47   16     3  607
#> 608       47   18     3  608
#> 609       47   20     3  609
#> 610       47   22     2  610
#> 611       47   24     2  611
#> 612       48    0     1  612
#> 613       48    2     1  613
#> 614       48    4     1  614
#> 615       48    6     1  615
#> 616       48    8     2  616
#> 617       48   10     2  617
#> 618       48   12     2  618
#> 619       48   14     2  619
#> 620       48   16     1  620
#> 621       48   18     1  621
#> 622       48   20     1  622
#> 623       48   22     1  623
#> 624       48   24     1  624
#> 625       49    0     1  625
#> 626       49    2     1  626
#> 627       49    4     1  627
#> 628       49    6     1  628
#> 629       49    8     2  629
#> 630       49   10     2  630
#> 631       49   12     2  631
#> 632       49   14     2  632
#> 633       49   16     2  633
#> 634       49   18     1  634
#> 635       49   20     1  635
#> 636       49   22     2  636
#> 637       49   24     2  637
#> 638       50    0     1  638
#> 639       50    2     1  639
#> 640       50    4     1  640
#> 641       50    6     2  641
#> 642       50    8     2  642
#> 643       50   10     2  643
#> 644       50   12     2  644
#> 645       50   14     2  645
#> 646       50   16     2  646
#> 647       50   18     2  647
#> 648       50   20     2  648
#> 649       50   22     2  649
#> 650       50   24     2  650
#> 651       51    0     1  651
#> 652       51    2     1  652
#> 653       51    4     1  653
#> 654       51    6     1  654
#> 655       51    8     1  655
#> 656       51   10     2  656
#> 657       51   12     2  657
#> 658       51   14     2  658
#> 659       51   16     2  659
#> 660       51   18     1  660
#> 661       51   20     1  661
#> 662       51   22     1  662
#> 663       51   24     1  663
#> 664       52    0     1  664
#> 665       52    2     1  665
#> 666       52    4     2  666
#> 667       52    6     2  667
#> 668       52    8     2  668
#> 669       52   10     2  669
#> 670       52   12     2  670
#> 671       52   14     2  671
#> 672       52   16     2  672
#> 673       52   18     2  673
#> 674       52   20     2  674
#> 675       52   22     3  675
#> 676       52   24     3  676
#> 677       53    0     1  677
#> 678       53    2     1  678
#> 679       53    4     1  679
#> 680       53    6     3  680
#> 681       53    8     3  681
#> 682       53   10     2  682
#> 683       53   12     3  683
#> 684       53   14     3  684
#> 685       53   16     2  685
#> 686       53   18     2  686
#> 687       53   20     3  687
#> 688       53   22     3  688
#> 689       53   24     3  689
#> 690       54    0     1  690
#> 691       54    2     2  691
#> 692       54    4     2  692
#> 693       54    6     2  693
#> 694       54    8     2  694
#> 695       54   10     2  695
#> 696       54   12     2  696
#> 697       54   14     2  697
#> 698       54   16     2  698
#> 699       54   18     2  699
#> 700       54   20     2  700
#> 701       54   22     2  701
#> 702       54   24     3  702
#> 703       55    0     1  703
#> 704       55    2     1  704
#> 705       55    4     1  705
#> 706       55    6     1  706
#> 707       55    8     1  707
#> 708       55   10     1  708
#> 709       55   12     1  709
#> 710       55   14     2  710
#> 711       55   16     2  711
#> 712       55   18     2  712
#> 713       55   20     2  713
#> 714       55   22     2  714
#> 715       55   24     2  715
#> 716       56    0     1  716
#> 717       56    2     1  717
#> 718       56    4     1  718
#> 719       56    6     1  719
#> 720       56    8     1  720
#> 721       56   10     1  721
#> 722       56   12     1  722
#> 723       56   14     1  723
#> 724       56   16     1  724
#> 725       56   18     1  725
#> 726       56   20     1  726
#> 727       56   22     1  727
#> 728       56   24     1  728
#> 729       57    0     1  729
#> 730       57    2     1  730
#> 731       57    4     2  731
#> 732       57    6     2  732
#> 733       57    8     2  733
#> 734       57   10     2  734
#> 735       57   12     1  735
#> 736       57   14     1  736
#> 737       57   16     1  737
#> 738       57   18     2  738
#> 739       57   20     2  739
#> 740       57   22     2  740
#> 741       57   24     2  741
#> 742       58    0     1  742
#> 743       58    2     1  743
#> 744       58    4     1  744
#> 745       58    6     1  745
#> 746       58    8     1  746
#> 747       58   10     1  747
#> 748       58   12     1  748
#> 749       58   14     1  749
#> 750       58   16     1  750
#> 751       58   18     1  751
#> 752       58   20     1  752
#> 753       58   22     2  753
#> 754       58   24     2  754
#> 755       59    0     1  755
#> 756       59    2     2  756
#> 757       59    4     2  757
#> 758       59    6     2  758
#> 759       59    8     1  759
#> 760       59   10     1  760
#> 761       59   12     1  761
#> 762       59   14     1  762
#> 763       59   16     1  763
#> 764       59   18     1  764
#> 765       59   20     1  765
#> 766       59   22     3  766
#> 767       59   24     3  767
#> 768       60    0     1  768
#> 769       60    2     1  769
#> 770       60    4     1  770
#> 771       60    6     2  771
#> 772       60    8     1  772
#> 773       60   10     1  773
#> 774       60   12     1  774
#> 775       60   14     1  775
#> 776       60   16     1  776
#> 777       60   18     1  777
#> 778       60   20     1  778
#> 779       60   22     1  779
#> 780       60   24     2  780
#> 781       61    0     1  781
#> 782       61    2     1  782
#> 783       61    4     1  783
#> 784       61    6     1  784
#> 785       61    8     1  785
#> 786       61   10     1  786
#> 787       61   12     1  787
#> 788       61   14     1  788
#> 789       61   16     1  789
#> 790       61   18     1  790
#> 791       61   20     1  791
#> 792       61   22     1  792
#> 793       61   24     1  793
#> 794       62    0     1  794
#> 795       62    2     1  795
#> 796       62    4     1  796
#> 797       62    6     2  797
#> 798       62    8     2  798
#> 799       62   10     1  799
#> 800       62   12     1  800
#> 801       62   14     1  801
#> 802       62   16     1  802
#> 803       62   18     2  803
#> 804       62   20     2  804
#> 805       62   22     3  805
#> 806       62   24     3  806
#> 807       63    0     1  807
#> 808       63    2     3  808
#> 809       63    4     3  809
#> 810       63    6     3  810
#> 811       63    8     3  811
#> 812       63   10     3  812
#> 813       63   12     3  813
#> 814       63   14     3  814
#> 815       63   16     3  815
#> 816       63   18     3  816
#> 817       63   20     3  817
#> 818       63   22     3  818
#> 819       63   24     3  819
#> 820       64    0     1  820
#> 821       64    2     2  821
#> 822       64    4     1  822
#> 823       64    6     1  823
#> 824       64    8     1  824
#> 825       64   10     1  825
#> 826       64   12     1  826
#> 827       64   14     1  827
#> 828       64   16     1  828
#> 829       64   18     2  829
#> 830       64   20     3  830
#> 831       64   22     3  831
#> 832       64   24     2  832
#> 833       65    0     1  833
#> 834       65    2     1  834
#> 835       65    4     1  835
#> 836       65    6     1  836
#> 837       65    8     1  837
#> 838       65   10     1  838
#> 839       65   12     1  839
#> 840       65   14     1  840
#> 841       65   16     1  841
#> 842       65   18     1  842
#> 843       65   20     2  843
#> 844       65   22     2  844
#> 845       65   24     2  845
#> 846       66    0     1  846
#> 847       66    2     1  847
#> 848       66    4     1  848
#> 849       66    6     1  849
#> 850       66    8     1  850
#> 851       66   10     1  851
#> 852       66   12     1  852
#> 853       66   14     1  853
#> 854       66   16     1  854
#> 855       66   18     1  855
#> 856       66   20     1  856
#> 857       66   22     1  857
#> 858       66   24     1  858
#> 859       67    0     1  859
#> 860       67    2     1  860
#> 861       67    4     1  861
#> 862       67    6     1  862
#> 863       67    8     1  863
#> 864       67   10     1  864
#> 865       67   12     2  865
#> 866       67   14     2  866
#> 867       67   16     2  867
#> 868       67   18     2  868
#> 869       67   20     2  869
#> 870       67   22     2  870
#> 871       67   24     2  871
#> 872       68    0     1  872
#> 873       68    2     1  873
#> 874       68    4     1  874
#> 875       68    6     2  875
#> 876       68    8     2  876
#> 877       68   10     2  877
#> 878       68   12     2  878
#> 879       68   14     2  879
#> 880       68   16     2  880
#> 881       68   18     2  881
#> 882       68   20     1  882
#> 883       68   22     1  883
#> 884       68   24     2  884
#> 885       69    0     1  885
#> 886       69    2     1  886
#> 887       69    4     1  887
#> 888       69    6     1  888
#> 889       69    8     1  889
#> 890       69   10     1  890
#> 891       69   12     2  891
#> 892       69   14     3  892
#> 893       69   16     3  893
#> 894       69   18     3  894
#> 895       69   20     3  895
#> 896       69   22     3  896
#> 897       69   24     3  897
#> 898       70    0     1  898
#> 899       70    2     1  899
#> 900       70    4     1  900
#> 901       70    6     2  901
#> 902       70    8     2  902
#> 903       70   10     2  903
#> 904       70   12     2  904
#> 905       70   14     2  905
#> 906       70   16     2  906
#> 907       70   18     2  907
#> 908       70   20     2  908
#> 909       70   22     2  909
#> 910       70   24     2  910
#> 911       71    0     1  911
#> 912       71    2     1  912
#> 913       71    4     1  913
#> 914       71    6     1  914
#> 915       71    8     1  915
#> 916       71   10     2  916
#> 917       71   12     2  917
#> 918       71   14     2  918
#> 919       71   16     2  919
#> 920       71   18     2  920
#> 921       71   20     2  921
#> 922       71   22     2  922
#> 923       71   24     2  923
#> 924       72    0     1  924
#> 925       72    2     1  925
#> 926       72    4     1  926
#> 927       72    6     3  927
#> 928       72    8     3  928
#> 929       72   10     3  929
#> 930       72   12     3  930
#> 931       72   14     1  931
#> 932       72   16     1  932
#> 933       72   18     2  933
#> 934       72   20     2  934
#> 935       72   22     3  935
#> 936       72   24     3  936
#> 937       73    0     1  937
#> 938       73    2     1  938
#> 939       73    4     2  939
#> 940       73    6     2  940
#> 941       73    8     2  941
#> 942       73   10     2  942
#> 943       73   12     2  943
#> 944       73   14     3  944
#> 945       73   16     3  945
#> 946       73   18     3  946
#> 947       73   20     1  947
#> 948       73   22     1  948
#> 949       73   24     1  949
#> 950       74    0     1  950
#> 951       74    2     1  951
#> 952       74    4     1  952
#> 953       74    6     1  953
#> 954       74    8     1  954
#> 955       74   10     1  955
#> 956       74   12     1  956
#> 957       74   14     1  957
#> 958       74   16     1  958
#> 959       74   18     2  959
#> 960       74   20     2  960
#> 961       74   22     2  961
#> 962       74   24     3  962
#> 963       75    0     1  963
#> 964       75    2     1  964
#> 965       75    4     1  965
#> 966       75    6     1  966
#> 967       75    8     1  967
#> 968       75   10     1  968
#> 969       75   12     1  969
#> 970       75   14     1  970
#> 971       75   16     1  971
#> 972       75   18     1  972
#> 973       75   20     1  973
#> 974       75   22     1  974
#> 975       75   24     2  975
#> 976       76    0     1  976
#> 977       76    2     1  977
#> 978       76    4     1  978
#> 979       76    6     1  979
#> 980       76    8     1  980
#> 981       76   10     2  981
#> 982       76   12     3  982
#> 983       76   14     3  983
#> 984       76   16     3  984
#> 985       76   18     2  985
#> 986       76   20     2  986
#> 987       76   22     2  987
#> 988       76   24     2  988
#> 989       77    0     1  989
#> 990       77    2     1  990
#> 991       77    4     1  991
#> 992       77    6     1  992
#> 993       77    8     1  993
#> 994       77   10     1  994
#> 995       77   12     1  995
#> 996       77   14     1  996
#> 997       77   16     1  997
#> 998       77   18     1  998
#> 999       77   20     1  999
#> 1000      77   22     2 1000
#> 1001      77   24     2 1001
#> 1002      78    0     1 1002
#> 1003      78    2     1 1003
#> 1004      78    4     1 1004
#> 1005      78    6     1 1005
#> 1006      78    8     1 1006
#> 1007      78   10     1 1007
#> 1008      78   12     2 1008
#> 1009      78   14     1 1009
#> 1010      78   16     2 1010
#> 1011      78   18     2 1011
#> 1012      78   20     2 1012
#> 1013      78   22     2 1013
#> 1014      78   24     2 1014
#> 1015      79    0     1 1015
#> 1016      79    2     1 1016
#> 1017      79    4     1 1017
#> 1018      79    6     1 1018
#> 1019      79    8     3 1019
#> 1020      79   10     3 1020
#> 1021      79   12     3 1021
#> 1022      79   14     3 1022
#> 1023      79   16     3 1023
#> 1024      79   18     3 1024
#> 1025      79   20     3 1025
#> 1026      79   22     3 1026
#> 1027      79   24     3 1027
#> 1028      80    0     1 1028
#> 1029      80    2     1 1029
#> 1030      80    4     1 1030
#> 1031      80    6     2 1031
#> 1032      80    8     2 1032
#> 1033      80   10     3 1033
#> 1034      80   12     3 1034
#> 1035      80   14     3 1035
#> 1036      80   16     3 1036
#> 1037      80   18     3 1037
#> 1038      80   20     3 1038
#> 1039      80   22     2 1039
#> 1040      80   24     2 1040
#> 1041      81    0     1 1041
#> 1042      81    2     1 1042
#> 1043      81    4     1 1043
#> 1044      81    6     2 1044
#> 1045      81    8     2 1045
#> 1046      81   10     2 1046
#> 1047      81   12     2 1047
#> 1048      81   14     2 1048
#> 1049      81   16     2 1049
#> 1050      81   18     3 1050
#> 1051      81   20     3 1051
#> 1052      81   22     3 1052
#> 1053      81   24     3 1053
#> 1054      82    0     1 1054
#> 1055      82    2     1 1055
#> 1056      82    4     2 1056
#> 1057      82    6     2 1057
#> 1058      82    8     2 1058
#> 1059      82   10     1 1059
#> 1060      82   12     1 1060
#> 1061      82   14     1 1061
#> 1062      82   16     2 1062
#> 1063      82   18     2 1063
#> 1064      82   20     2 1064
#> 1065      82   22     2 1065
#> 1066      82   24     2 1066
#> 1067      83    0     1 1067
#> 1068      83    2     1 1068
#> 1069      83    4     2 1069
#> 1070      83    6     2 1070
#> 1071      83    8     2 1071
#> 1072      83   10     2 1072
#> 1073      83   12     2 1073
#> 1074      83   14     2 1074
#> 1075      83   16     2 1075
#> 1076      83   18     2 1076
#> 1077      83   20     2 1077
#> 1078      83   22     2 1078
#> 1079      83   24     2 1079
#> 1080      84    0     1 1080
#> 1081      84    2     1 1081
#> 1082      84    4     1 1082
#> 1083      84    6     1 1083
#> 1084      84    8     1 1084
#> 1085      84   10     1 1085
#> 1086      84   12     1 1086
#> 1087      84   14     1 1087
#> 1088      84   16     1 1088
#> 1089      84   18     1 1089
#> 1090      84   20     1 1090
#> 1091      84   22     1 1091
#> 1092      84   24     2 1092
#> 1093      85    0     1 1093
#> 1094      85    2     1 1094
#> 1095      85    4     1 1095
#> 1096      85    6     1 1096
#> 1097      85    8     1 1097
#> 1098      85   10     1 1098
#> 1099      85   12     1 1099
#> 1100      85   14     1 1100
#> 1101      85   16     2 1101
#> 1102      85   18     2 1102
#> 1103      85   20     2 1103
#> 1104      85   22     2 1104
#> 1105      85   24     2 1105
#> 1106      86    0     1 1106
#> 1107      86    2     2 1107
#> 1108      86    4     2 1108
#> 1109      86    6     3 1109
#> 1110      86    8     3 1110
#> 1111      86   10     3 1111
#> 1112      86   12     3 1112
#> 1113      86   14     3 1113
#> 1114      86   16     3 1114
#> 1115      86   18     3 1115
#> 1116      86   20     3 1116
#> 1117      86   22     3 1117
#> 1118      86   24     3 1118
#> 1119      87    0     1 1119
#> 1120      87    2     1 1120
#> 1121      87    4     1 1121
#> 1122      87    6     1 1122
#> 1123      87    8     1 1123
#> 1124      87   10     1 1124
#> 1125      87   12     1 1125
#> 1126      87   14     1 1126
#> 1127      87   16     1 1127
#> 1128      87   18     2 1128
#> 1129      87   20     2 1129
#> 1130      87   22     2 1130
#> 1131      87   24     2 1131
#> 1132      88    0     1 1132
#> 1133      88    2     1 1133
#> 1134      88    4     1 1134
#> 1135      88    6     2 1135
#> 1136      88    8     2 1136
#> 1137      88   10     2 1137
#> 1138      88   12     2 1138
#> 1139      88   14     2 1139
#> 1140      88   16     2 1140
#> 1141      88   18     2 1141
#> 1142      88   20     2 1142
#> 1143      88   22     2 1143
#> 1144      88   24     3 1144
#> 1145      89    0     1 1145
#> 1146      89    2     1 1146
#> 1147      89    4     2 1147
#> 1148      89    6     2 1148
#> 1149      89    8     2 1149
#> 1150      89   10     2 1150
#> 1151      89   12     3 1151
#> 1152      89   14     3 1152
#> 1153      89   16     3 1153
#> 1154      89   18     3 1154
#> 1155      89   20     3 1155
#> 1156      89   22     3 1156
#> 1157      89   24     3 1157
#> 1158      90    0     1 1158
#> 1159      90    2     1 1159
#> 1160      90    4     1 1160
#> 1161      90    6     2 1161
#> 1162      90    8     2 1162
#> 1163      90   10     3 1163
#> 1164      90   12     3 1164
#> 1165      90   14     1 1165
#> 1166      90   16     1 1166
#> 1167      90   18     1 1167
#> 1168      90   20     1 1168
#> 1169      90   22     2 1169
#> 1170      90   24     2 1170
#> 1171      91    0     1 1171
#> 1172      91    2     1 1172
#> 1173      91    4     1 1173
#> 1174      91    6     1 1174
#> 1175      91    8     1 1175
#> 1176      91   10     1 1176
#> 1177      91   12     1 1177
#> 1178      91   14     2 1178
#> 1179      91   16     2 1179
#> 1180      91   18     2 1180
#> 1181      91   20     2 1181
#> 1182      91   22     2 1182
#> 1183      91   24     2 1183
#> 1184      92    0     1 1184
#> 1185      92    2     1 1185
#> 1186      92    4     3 1186
#> 1187      92    6     3 1187
#> 1188      92    8     3 1188
#> 1189      92   10     3 1189
#> 1190      92   12     3 1190
#> 1191      92   14     2 1191
#> 1192      92   16     2 1192
#> 1193      92   18     2 1193
#> 1194      92   20     2 1194
#> 1195      92   22     2 1195
#> 1196      92   24     2 1196
#> 1197      93    0     1 1197
#> 1198      93    2     1 1198
#> 1199      93    4     1 1199
#> 1200      93    6     1 1200
#> 1201      93    8     1 1201
#> 1202      93   10     1 1202
#> 1203      93   12     1 1203
#> 1204      93   14     1 1204
#> 1205      93   16     1 1205
#> 1206      93   18     2 1206
#> 1207      93   20     1 1207
#> 1208      93   22     1 1208
#> 1209      93   24     1 1209
#> 1210      94    0     1 1210
#> 1211      94    2     1 1211
#> 1212      94    4     1 1212
#> 1213      94    6     1 1213
#> 1214      94    8     1 1214
#> 1215      94   10     1 1215
#> 1216      94   12     1 1216
#> 1217      94   14     1 1217
#> 1218      94   16     1 1218
#> 1219      94   18     1 1219
#> 1220      94   20     1 1220
#> 1221      94   22     1 1221
#> 1222      94   24     2 1222
#> 1223      95    0     1 1223
#> 1224      95    2     1 1224
#> 1225      95    4     1 1225
#> 1226      95    6     1 1226
#> 1227      95    8     1 1227
#> 1228      95   10     1 1228
#> 1229      95   12     2 1229
#> 1230      95   14     2 1230
#> 1231      95   16     2 1231
#> 1232      95   18     2 1232
#> 1233      95   20     2 1233
#> 1234      95   22     2 1234
#> 1235      95   24     1 1235
#> 1236      96    0     1 1236
#> 1237      96    2     1 1237
#> 1238      96    4     1 1238
#> 1239      96    6     2 1239
#> 1240      96    8     2 1240
#> 1241      96   10     2 1241
#> 1242      96   12     2 1242
#> 1243      96   14     2 1243
#> 1244      96   16     3 1244
#> 1245      96   18     3 1245
#> 1246      96   20     3 1246
#> 1247      96   22     3 1247
#> 1248      96   24     3 1248
#> 1249      97    0     1 1249
#> 1250      97    2     1 1250
#> 1251      97    4     1 1251
#> 1252      97    6     1 1252
#> 1253      97    8     2 1253
#> 1254      97   10     2 1254
#> 1255      97   12     2 1255
#> 1256      97   14     2 1256
#> 1257      97   16     2 1257
#> 1258      97   18     3 1258
#> 1259      97   20     3 1259
#> 1260      97   22     3 1260
#> 1261      97   24     3 1261
#> 1262      98    0     1 1262
#> 1263      98    2     2 1263
#> 1264      98    4     2 1264
#> 1265      98    6     2 1265
#> 1266      98    8     3 1266
#> 1267      98   10     3 1267
#> 1268      98   12     3 1268
#> 1269      98   14     3 1269
#> 1270      98   16     3 1270
#> 1271      98   18     3 1271
#> 1272      98   20     3 1272
#> 1273      98   22     3 1273
#> 1274      98   24     2 1274
#> 1275      99    0     1 1275
#> 1276      99    2     1 1276
#> 1277      99    4     2 1277
#> 1278      99    6     3 1278
#> 1279      99    8     3 1279
#> 1280      99   10     2 1280
#> 1281      99   12     2 1281
#> 1282      99   14     2 1282
#> 1283      99   16     2 1283
#> 1284      99   18     2 1284
#> 1285      99   20     2 1285
#> 1286      99   22     2 1286
#> 1287      99   24     2 1287
#> 1288     100    0     1 1288
#> 1289     100    2     2 1289
#> 1290     100    4     1 1290
#> 1291     100    6     1 1291
#> 1292     100    8     1 1292
#> 1293     100   10     1 1293
#> 1294     100   12     1 1294
#> 1295     100   14     2 1295
#> 1296     100   16     2 1296
#> 1297     100   18     2 1297
#> 1298     100   20     2 1298
#> 1299     100   22     2 1299
#> 1300     100   24     2 1300