Update the maximum likelihood estimates in a fitted model object.

Update the maximum likelihood estimates in a fitted model object. Intended for developer use only.

Usage

updatepars.msm(x, pars)

Arguments

x

A fitted multi-state model object, as returned by msm.

pars

Vector of new parameters, in their untransformed real-line parameterisations, to substitute for the maximum likelihood estimates corresponding to those in the estimates component of the fitted model object (msm.object). The order of the parameters is documented in msm, argument fixedpars.

Value

An updated msm model object with the updated maximum likelihood estimates, but with the covariances / standard errors unchanged.

Point estimates from output functions such as qmatrix.msm, pmatrix.msm, or any related function, can then be evaluated with the new parameters, and at arbitrary covariate values.

This function is used, for example, when computing confidence intervals from pmatrix.msm, and related functions, using the ci="normal" method.

Author

C. H. Jackson chris.jackson@mrc-bsu.cam.ac.uk